Document fallback with example styles
As discussed on #71, there's not much the component can do itself to provide a fallback, but it's very valuable to document how the fallback works for people less familiar with web components.

This adds a prominent example and educational info that will hopefully encourage people to provide a fallback for the numerous instances when it might be needed. It includes some [opinionated example styles](https://codepen.io/mrwweb/pen/qBKQOPW) which specifically call out using `aspect-ratio` to match the normal dimensions of a YouTube embed.

Adding this documentation feels like enough to satisfy #71 in my opinion, but I'd like to hear @simevidas weigh in on that too.

Additional minor change: remove whitespace at start of lines in playlist example that no other example code has.